% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{chisquare}
\alias{chisquare}
\title{Calculate Chisquare}
\usage{
chisquare(d, D, S, P = NULL, cholZ = NULL)
}
\arguments{
\item{D}{TODO}

\item{S}{TODO}

\item{P}{TODO}

\item{expdata}{TODO}

\item{moddata}{TODO}
}
\value{
TODO
}
\description{
Calculate Chisquare
}
