% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{createMLOptimFuns}
\alias{createMLOptimFuns}
\title{Create Functions for ML Optimization}
\usage{
createMLOptimFuns()
}
\value{
List with functions \code{logLike} and \code{gradLogLike}
}
\description{
Create Functions for ML Optimization
}
