% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysCompHandler_GP_sqrexp.R
\name{createSysCompGPHandler}
\alias{createSysCompGPHandler}
\title{Create Handler for Systematic GP Components}
\usage{
createSysCompGPHandler()
}
\value{
TODO
}
\description{
Create Handler for Systematic GP Components
}
