% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysCompHandler_norm.R
\name{createSysCompNormHandler}
\alias{createSysCompNormHandler}
\title{Create Handler for Systematic Normalization Components}
\usage{
createSysCompNormHandler(dataref = NULL)
}
\value{
TODO
}
\description{
Create Handler for Systematic Normalization Components
}
