% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysCompHandler_reac.R
\name{createSysCompReacHandler}
\alias{createSysCompReacHandler}
\title{Create Systematic Component Handler for Reactions}
\usage{
createSysCompReacHandler(subentList)
}
\arguments{
\item{subentList}{TODO}
}
\value{
TODO
}
\description{
Create Systematic Component Handler for Reactions
}
