% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicUtilityFuns.R
\name{getDt_setDt_x}
\alias{getDt_setDt_x}
\alias{setDt_DATA}
\alias{getDt_DATA}
\alias{getDt_REFDATA}
\alias{setDt_UNC}
\alias{getDt_UNC}
\title{Convenience Functions to Set and Get Columns in a Datatable}
\usage{
setDt_DATA(dt, x, idx = NULL)

getDt_DATA(dt, idx = NULL)

getDt_REFDATA(dt, idx = NULL)

setDt_UNC(dt, x, idx = NULL)

getDt_UNC(dt, idx = NULL)
}
\arguments{
\item{dt}{a datatable}

\item{x}{a vector with new values of the column}

\item{idx}{if \code{NULL}, \code{x} will replace the complete column, otherwise
only the values associated with the indices in \code{idx} will be changed.}
}
\note{
Datatables are modified by reference, i.e. the datatable \code{dt} passed as argument is modified!
}
