% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianUpdateFuns.R
\name{getPosteriorSys}
\alias{getPosteriorSys}
\title{Get Posterior of Systematic Components}
\usage{
getPosteriorSys(expDt, sysDt, gpDt = NULL, sysCompHandler)
}
\arguments{
\item{expDt}{datatable with experimental data}

\item{sysDt}{datatable with systematic components}

\item{gpDt}{datatable with Gaussian process components}

\item{sysCompHandler}{systematic component handler, see \link{createSysCompHandler}}
}
\value{
list with two elements \code{sysDt} and \code{cov} which contain
the posterior of the systematic components and the covariance matrix, respectively.
and covariance matrix
}
\description{
Get Posterior of Systematic Components
}
