% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{gradLogLike}
\alias{gradLogLike}
\title{Calculate Gradient of Log-Likelihood}
\usage{
gradLogLike(d, D, S, cholZ = NULL, sysDt, gpDt = NULL, statUncIdx = NULL,
  sysUncIdx = NULL, gpIdx = NULL, sysCompHandler)
}
\arguments{
\item{D}{TODO}

\item{S}{TODO}

\item{sysDt}{TODO}

\item{sysCompHandler}{TODO}

\item{expdata}{TODO}

\item{moddata}{TODO}

\item{idxDt}{TODO}
}
\value{
TODO
}
\description{
Calculate Gradient of Log-Likelihood
}
