% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicUtilityFuns.R
\name{isValidSysDt}
\alias{isValidSysDt}
\title{Verify Structure of SysDt}
\usage{
isValidSysDt(sysDt, range.check = TRUE, print.info = TRUE)
}
\arguments{
\item{sysDt}{TODO}

\item{print.info}{TODO}

\item{before.prep}{TODO}
}
\value{
TODO
}
\description{
Verify Structure of SysDt
}
