% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMalgo.R
\name{LMalgo}
\alias{LMalgo}
\title{Levenberg-Marquardt algorithm with prior}
\usage{
LMalgo(fn, jac, pinit, p0, P0, yexp, D, S, X, lower = -Inf, upper = Inf,
  logger = NULL, control = list(maxit = 10, abstol = 0.001, reltol = 0.001,
  mu = NULL))
}
\arguments{
\item{fn}{model function}

\item{jac}{Jacobian of model function}

\item{pinit}{initial model parameters}

\item{p0}{prior mean of model parameters}

\item{P0}{prior covariance matrix of model parameters}

\item{yexp}{experimental data points}

\item{D}{only diagonal contribution to experimental covariance matrix}

\item{S}{mapping from systematic components to observations}

\item{X}{prior covariance matrix for systematic components}

\item{control}{control arguments, see \code{details}}
}
\value{
list with elements \code{par}, \code{fn}, \code{jac}, \code{counts}, \code{value}
}
\description{
Levenberg-Marquardt algorithm with prior
}
\details{
The control argument is a list that can contain
\itemize{
  \item maximum number of iterations \code{maxit}
  \item absolute tolerance required for convergence \code{abstol}
  \item relative tolerance required for convergence \code{reltol}
}
}
