% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{logDetCov}
\alias{logDetCov}
\title{Logarithm of Determinant of Covariance matrix}
\usage{
logDetCov(D, S, P, cholZ = NULL)
}
\arguments{
\item{D}{TODO}

\item{S}{TODO}

\item{P}{TODO}
}
\value{
TODO
}
\description{
Logarithm of Determinant of Covariance matrix
}
