% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{logLike}
\alias{logLike}
\title{Calculate Log-Likelihood}
\usage{
logLike(d, D, S, P, cholZ = NULL)
}
\arguments{
\item{D}{TODO}

\item{S}{TODO}

\item{P}{TODO}

\item{expdata}{TODO}

\item{moddata}{TODO}
}
\value{
TODO
}
\description{
Calculate Log-Likelihood
}
