% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianUpdateFuns.R
\name{mapSysToExp}
\alias{mapSysToExp}
\title{Map Systematic Components to Experimental Grid}
\usage{
mapSysToExp(expDt, sysList, sysCompHandler, ret.cov = FALSE, idx = NULL)
}
\arguments{
\item{expDt}{datatable with experimental data}

\item{sysList}{list with components \code{sysDt} and potentially \code{cov}}

\item{sysCompHandler}{handler for systematic components as created by \link{createSysCompHandler}}

\item{ret.cov}{should the posterior matrix be computed?}

\item{idx}{vector of indices that specifies which systematic components should be mapped.
Default \code{NULL} means to map all of them.}
}
\value{
list with two components \code{expDt} and \code{cov}
}
\description{
Map Systematic Components to Experimental Grid
}
