% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{mult_xt_invCov_x}
\alias{mult_xt_invCov_x}
\title{Multiply an Inverse Covariance Matrix from Both Sides}
\usage{
mult_xt_invCov_x(x, D, S, P = NULL, cholZ = NULL)
}
\arguments{
\item{x}{TODO}

\item{D}{TODO}

\item{S}{TODO}

\item{P}{TODO}
}
\value{
TODO
}
\description{
Multiply an Inverse Covariance Matrix from Both Sides
}
