% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmap_piecewise.R
\name{pwMap}
\alias{pwMap}
\title{Create Piecewise Linear Map}
\usage{
pwMap(expEn, opts = NULL, ret.mat = TRUE)
}
\arguments{
\item{expEn}{TODO}

\item{opts}{TODO}

\item{ret.mat}{TODO}
}
\value{
TODO
}
\description{
Create Piecewise Linear Map
}
