% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marlikefuns.R
\name{tr_invCov_SdPtS}
\alias{tr_invCov_SdPtS}
\title{Trace of Inverse Covariance Matrix multplied with derivative of P}
\usage{
tr_invCov_SdPtS(dP, D, S, P = NULL, cholZ = NULL)
}
\arguments{
\item{dP}{TODO}

\item{D}{TODO}

\item{S}{TODO}

\item{P}{TODO}

\item{cholZ}{TODO}
}
\value{
TODO
}
\description{
Trace of Inverse Covariance Matrix multplied with derivative of P
}
